build(poetry.lock): update certifi and charset-normalizer packages to newer versions The certifi package was updated from version 2023.7.22 to 2024.2.2, and the charset-normalizer package was updated from version 3.2.0 to 3.3.2. These updates ensure that the project uses the latest versions of these packages, potentially including bug fixes, security patches, and new features.
chore: update wheel files for charset_normalizer-3.2.0 for various platforms
The commit updates the wheel files for charset_normalizer-3.2.0 for different platforms to ensure compatibility and provide the latest version of the library for users on those platforms.

feat: update charset_normalizer package to version 3.3.2
The commit introduces an update to the charset_normalizer package, bumping it to version 3.3.2. This update includes new wheel files for various platforms and architectures, each with their respective hash values. The update ensures that the latest features and improvements from the new version are available for use in the project.

feat: add new wheel files for charset_normalizer 3.3.2 for various platforms
The commit adds new wheel files for charset_normalizer 3.3.2 for different platforms such as manylinux, musllinux, macOS, and Windows. These wheel files provide compatibility for users running the application on different operating systems and architectures.

feat(dependencies): update versions of charset_normalizer, idna, and requests packages
The commit updates the versions of the charset_normalizer, idna, and requests packages to their latest versions. This ensures that the project uses the most recent features, improvements, and bug fixes provided by the updated packages.

fix(poetry.lock): update urllib3 package version to 2.2.1 and adjust python-versions to ">=3.8"
The urllib3 package version has been updated to 2.2.1 in order to incorporate the latest changes and improvements. Additionally, the python-versions constraint has been adjusted to ">=3.8" to reflect the compatibility requirements for the updated urllib3 version.