Refactor/major rewrite (#19)
* build(poetry.lock): update certifi and charset-normalizer packages to newer versions
The certifi package was updated from version 2023.7.22 to 2024.2.2, and the charset-normalizer package was updated from version 3.2.0 to 3.3.2. These updates ensure that the project uses the latest versions of these packages, potentially including bug fixes, security patches, and new features.

chore: update wheel files for charset_normalizer-3.2.0 for various platforms
The commit updates the wheel files for charset_normalizer-3.2.0 for different platforms to ensure compatibility and provide the latest version of the library for users on those platforms.

feat: update charset_normalizer package to version 3.3.2
The commit introduces an update to the charset_normalizer package, bumping it to version 3.3.2. This update includes new wheel files for various platforms and architectures, each with their respective hash values. The update ensures that the latest features and improvements from the new version are available for use in the project.

feat: add new wheel files for charset_normalizer 3.3.2 for various platforms
The commit adds new wheel files for charset_normalizer 3.3.2 for different platforms such as manylinux, musllinux, macOS, and Windows. These wheel files provide compatibility for users running the application on different operating systems and architectures.

feat(dependencies): update versions of charset_normalizer, idna, and requests packages
The commit updates the versions of the charset_normalizer, idna, and requests packages to their latest versions. This ensures that the project uses the most recent features, improvements, and bug fixes provided by the updated packages.

fix(poetry.lock): update urllib3 package version to 2.2.1 and adjust python-versions to ">=3.8"
The urllib3 package version has been updated to 2.2.1 in order to incorporate the latest changes and improvements. Additionally, the python-versions constraint has been adjusted to ">=3.8" to reflect the compatibility requirements for the updated urllib3 version.

* ci(codeql-analysis.yml): update cron schedule for codeql-analysis workflow
ci(create-new-release.yml): create workflow for generating new releases and release notes
ci(dependency-review.yml): update dependency-review-action to version 4
ci(generate-requirements.yml): add workflow for generating requirements files based on changes in pyproject.toml and poetry.lock
ci(mypy-linting.yml): add mypy linting workflow for Python codebase
ci(pytests-coverage.yml): add workflow for running PyTests and generating coverage report

The changes were made to enhance the CI/CD workflows for the project. The updates include adjusting cron schedules, adding new workflows for release generation, dependency review, requirements file generation, mypy linting, and PyTests with coverage reporting. These changes aim to improve automation, code quality, and release management processes.

* chore(.gitignore): update .gitignore file to include more patterns
docs(CONTRIBUTING.md): add contributing guidelines
docs(README.md): update README with detailed usage instructions
The .gitignore file was updated to include more patterns, which will help to keep the repository clean by ignoring files that should not be tracked. A CONTRIBUTING.md file was added to provide guidelines for contributors, which will help to maintain a consistent codebase and streamline the contribution process. The README was updated with detailed usage instructions to make it easier for users to understand how to use the library.

* feat: Update project dependencies and metadata in pyproject.toml and requirements.txt
The project dependencies have been updated to their latest versions to ensure compatibility with the latest features and security updates. The project metadata in pyproject.toml has also been updated to reflect the new version and a more accurate description of the project. A new dev-requirements.txt file has been added to separate development dependencies from production ones, making it easier to manage dependencies in different environments.

* refactor(trackerstatus): restructure code for better modularity and readability
The previous structure of the trackerstatus module was monolithic and difficult to maintain. The code has been refactored into separate modules and classes for better modularity and readability. The APIClient class has been introduced to handle API requests, and separate endpoint classes (BTNEndpoint and StatusEndpoint) have been created to handle specific API endpoints. This makes the code easier to understand, maintain, and extend in the future.

* test: add unit tests for btn and status endpoints
The new tests ensure that the btn and status endpoints are working as expected. They check the structure and some key values of the responses. This will help to catch any regressions or breaking changes in the future.

* ci(mypy-linting.yml): change mypy target from all python files to specific directories
The mypy linting target has been changed from all Python files in the repository to specifically the 'trackerstatus' and 'tests' directories. This change was made to focus linting efforts on the most relevant parts of the codebase and to avoid unnecessary linting of files that are not part of the main application or test suite.

* Update README.md

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>

* Update tests/test_btn.py

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>

* Update tests/test_btn.py

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>

* Update tests/test_status.py

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>

* Update trackerstatus/core.py

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>

---------

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>